IAZIS_lab4/requirements.txt
# Основные зависимости
requests>=2.31.0
python-dotenv>=1.0.0
nltk>=3.8.0

# GUI и интерфейс
tkinter  # обычно встроен, но упоминаем для ясности

# Дополнительные зависимости для улучшения функционала
Pillow>=10.0.0  # для работы с изображениями (если понадобится)
pandas>=2.0.0  # для расширенной обработки данных
matplotlib>=3.7.0  # для визуализации (опционально)